misc/HtmlTestSuite/results/test-1098.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml" lang="en">
<head>
  <title>Tokenizer states</title>
<link href="mailto:hsivonen@hsivonen.fi" rel="author" title="Henri Sivonen"/>
<link href="https://html.spec.whatwg.org/#html-fragment-parsing-algorithm" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head>
<body>
<script>
test(function() {
  var e = document.createElement("title");
  e.innerHTML = "&lt;/title&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&amp;lt;/title&amp;gt;&amp;lt;div&amp;gt;");
}, "&lt;/title&gt; should not break out of title.");

test(function() {
  var e = document.createElement("textarea");
  e.innerHTML = "&lt;/textarea&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&amp;lt;/textarea&amp;gt;&amp;lt;div&amp;gt;");
}, "&lt;/textarea&gt; should not break out of textarea.");

test(function() {
  var e = document.createElement("style");
  e.innerHTML = "&lt;/style&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/style&gt;&lt;div&gt;");
}, "&lt;/style&gt; should not break out of style.");

test(function() {
  var e = document.createElement("xmp");
  e.innerHTML = "&lt;/xmp&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/xmp&gt;&lt;div&gt;");
}, "&lt;/xmp&gt; should not break out of xmp.");

test(function() {
  var e = document.createElement("iframe");
  e.innerHTML = "&lt;/iframe&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/iframe&gt;&lt;div&gt;");
}, "&lt;/iframe&gt; should not break out of iframe.");

test(function() {
  var e = document.createElement("noembed");
  e.innerHTML = "&lt;/noembed&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/noembed&gt;&lt;div&gt;");
}, "&lt;/noembed&gt; should not break out of noembed.");

test(function() {
  var e = document.createElement("noframes");
  e.innerHTML = "&lt;/noframes&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/noframes&gt;&lt;div&gt;");
}, "&lt;/noframes&gt; should not break out of noframes.");

test(function() {
  var e = document.createElement("script");
  e.innerHTML = "&lt;\/script&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;\/script&gt;&lt;div&gt;");
}, "&lt;\/script&gt; should not break out of script.");


test(function() {
  var e = document.createElement("noscript");
  e.innerHTML = "&lt;/noscript&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/noscript&gt;&lt;div&gt;");
}, "&lt;/noscript&gt; should not break out of noscript.");

test(function() {
  var e = document.createElement("plaintext");
  e.innerHTML = "&lt;/plaintext&gt;&lt;div&gt;";
  assert_equals(e.getElementsByTagName("div").length, 0);
  assert_equals(e.innerHTML, "&lt;/plaintext&gt;&lt;div&gt;");
}, "&lt;/plaintext&gt; should not break out of plaintext.");

</script>
</body>
</html>
